Add files via upload
In this study, we investigate the factors that may influence whether or not a football (soccer) manager gets sacked. To achieve this, we collected data on various performance indicators, such as league position, number of wins, losses, draws, goals scored and goals conceded, along with managerial information such as the number of games overseen, winning, and losing streaks, and the date when the manager was sacked.

The study also looks at recent examples of managerial dismissals in professional football, including those involving Thomas Tuchel at Chelsea, Antonio Conte at Tottenham, and Liverpool's decision to retain their manager despite a recent dip in form. The time frame, we are taking into account is the previous six seasons, which allows for a more comprehensive analysis of trends and patterns and their consequences that may emerge.

Our study utilizes different software tools such as Microsoft Excel and RapidMiner to analyze the data and employ machine learning algorithms. We apply classification and clustering techniques to identify patterns in the data and determine which factors are most strongly associated with the likelihood of a manager being sacked. Our results suggest that the number of games managed, league position, and winning and losing streaks are the most significant factors, while sentiment analysis may provide additional insights into the decision-making processes of club management.

By providing insights into the factors that contribute to managerial dismissals and their impact on club performance, the study offers a valuable contribution to the world of professional football.